+++ src/interop/mod.rs
@@ -1,4 +1,5 @@
 pub mod async_messaging_helper;
 pub mod bridge;
 pub mod custom_highlighter;
+pub mod language;
 pub mod recommendation;